Move tests to appropriate folders
