+++ src/StaticAnimator/__tests__/StaticAnimator.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import { View } from 'react-native';
 
-import StaticAnimator from '../src/StaticAnimator';
+import StaticAnimator from '../';
 
 // FIXME: shouldAnimateOut, delay, duration, shouldRepeat and shouldLoop were occluded from these tests
 // due to the async complexity they introduce
